WIP moving task scheduler queries into scheduler package
